refactor : 패키지 구조 단순화
