--- src/main/java/gift/GlobalExceptionHandler.java
@@ -1,51 +1,0 @@
-package gift;
-
-import gift.member.error.ForbiddenException;
-import gift.member.error.UnauthorizedException;
-import gift.product.error.AlreadyExistsException;
-import gift.product.error.NotFoundException;
-import java.util.HashMap;
-import java.util.Map;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(NotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public String handleNotFoundException(NotFoundException ex, Model model) {
-        model.addAttribute("errorMessage", ex.getMessage());
-        return "error";
-    }
-
-    @ExceptionHandler(AlreadyExistsException.class)
-    @ResponseStatus(HttpStatus.CONFLICT)
-    public String handleAlreadyExistsException(AlreadyExistsException ex, Model model) {
-        model.addAttribute("errorMessage", ex.getMessage());
-        return "error";
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public String handleValidationException(MethodArgumentNotValidException ex, Model model){
-        model.addAttribute("errorMessage", "유효성 검사 실패: " + ex.getBindingResult().getFieldError().getDefaultMessage());
-        return "error";
-    }
-
-    @ExceptionHandler(UnauthorizedException.class)
-    public ResponseEntity<String> handleUnauthorizedException(UnauthorizedException ex) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-            .header("WWW-Authenticate", "Bearer realm=\"example\"")
-            .body(ex.getMessage());
-    }
-
-    @ExceptionHandler(ForbiddenException.class)
-    public ResponseEntity<String> handleForbiddenException(ForbiddenException ex) {
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ex.getMessage());
-    }
-}
