+++ src/main/java/gift/domain/Member.java
@@ -1,0 +1,50 @@
+package gift.domain;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+
+public class Member {
+
+    private Long id;
+
+    @NotNull(message = "email을 작성해주세요")
+    @Email
+    private String email;
+    @Size(min = 4, max = 20, message = "비밀번호는 4~20자 사이로 작성해주세요")
+    private String password;
+
+    public Member() {
+
+    }
+
+    public Member(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+}
