+++ src/main/java/gift/domain/Token.java
@@ -1,0 +1,15 @@
+package gift.domain;
+
+public class Token {
+
+    private String token;
+
+    public Token(String token) {
+        this.token = token;
+    }
+
+    public String getTokenDTO() {
+        return token;
+    }
+
+}
