+++ src/main/java/gift/error/AlreadyExistsException.java
@@ -1,0 +1,7 @@
+package gift.error;
+
+public class AlreadyExistsException extends RuntimeException{
+    public AlreadyExistsException(String message) {
+        super(message);
+    }
+}
