+++ src/main/java/gift/error/ForbiddenException.java
@@ -1,0 +1,7 @@
+package gift.error;
+
+public class ForbiddenException extends RuntimeException {
+    public ForbiddenException(String message) {
+        super(message);
+    }
+}
