+++ src/main/java/gift/error/NotFoundException.java
@@ -1,0 +1,7 @@
+package gift.error;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
