+++ src/main/java/gift/error/UnauthorizedException.java
@@ -1,0 +1,7 @@
+package gift.error;
+
+public class UnauthorizedException extends RuntimeException {
+    public UnauthorizedException(String message) {
+        super(message);
+    }
+}
