--- src/main/java/gift/member/config/WebConfig.java
@@ -1,23 +1,0 @@
-package gift.member.config;
-
-import gift.member.interceptor.JwtInterceptor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    private JwtInterceptor jwtInterceptor;
-
-    public WebConfig(JwtInterceptor jwtInterceptor) {
-        this.jwtInterceptor = jwtInterceptor;
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(jwtInterceptor)
-            .addPathPatterns("/**") // 모든 요청에 대해 인터셉터를 적용
-            .excludePathPatterns("/members/register", "/members/login"); // 로그인 및 회원가입 요청은 제외
-    }
-}
