--- src/main/java/gift/member/controller/MemberController.java
@@ -1,34 +1,0 @@
-package gift.member.controller;
-
-import gift.member.domain.Token;
-import gift.member.domain.Member;
-import gift.member.service.MemberService;
-import jakarta.validation.Valid;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/members")
-public class MemberController {
-    private MemberService memberService;
-
-    public MemberController(MemberService memberService) {
-        this.memberService = memberService;
-    }
-
-    @PostMapping("/register")
-    public ResponseEntity<?> register(@Valid @RequestBody Member member) {
-        Token token = memberService.register(member);
-        return ResponseEntity.ok(token);
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<?> login(@Valid @RequestBody Member member) {
-        Token token = memberService.login(member);
-        return ResponseEntity.ok(token);
-    }
-
-}
