--- src/main/java/gift/member/domain/Token.java
@@ -1,15 +1,0 @@
-package gift.member.domain;
-
-public class Token {
-
-    private String token;
-
-    public Token(String token) {
-        this.token = token;
-    }
-
-    public String getTokenDTO() {
-        return token;
-    }
-
-}
