--- src/main/java/gift/member/error/ForbiddenException.java
@@ -1,7 +1,0 @@
-package gift.member.error;
-
-public class ForbiddenException extends RuntimeException {
-    public ForbiddenException(String message) {
-        super(message);
-    }
-}
