--- src/main/java/gift/member/error/UnauthorizedException.java
@@ -1,7 +1,0 @@
-package gift.member.error;
-
-public class UnauthorizedException extends RuntimeException {
-    public UnauthorizedException(String message) {
-        super(message);
-    }
-}
