--- src/main/java/gift/member/repository/MemberRepository.java
@@ -1,13 +1,0 @@
-package gift.member.repository;
-
-import gift.member.domain.Member;
-import gift.member.domain.MemberEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
-    MemberEntity findByEmail(String email);
-}
