--- src/main/java/gift/member/util/PasswordUtil.java
@@ -1,14 +1,0 @@
-package gift.member.util;
-
-import java.util.Base64;
-
-public class PasswordUtil {
-
-    public static String encodePassword(String password) {
-        return Base64.getEncoder().encodeToString(password.getBytes());
-    }
-
-    public static String decodePassword(String encodedPassword) {
-        return new String(Base64.getDecoder().decode(encodedPassword));
-    }
-}
