--- src/main/java/gift/product/error/AlreadyExistsException.java
@@ -1,7 +1,0 @@
-package gift.product.error;
-
-public class AlreadyExistsException extends RuntimeException{
-    public AlreadyExistsException(String message) {
-        super(message);
-    }
-}
