--- src/main/java/gift/product/error/NotFoundException.java
@@ -1,7 +1,0 @@
-package gift.product.error;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
