--- src/main/java/gift/product/repository/ProductRepository.java
@@ -1,11 +1,0 @@
-package gift.product.repository;
-
-import gift.product.domain.ProductEntity;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
-    List<ProductEntity> findByNameContaining(String name);
-}
