+++ src/main/java/gift/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package gift.repository;
+
+import gift.entity.MemberEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
+    MemberEntity findByEmail(String email);
+}
