+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,0 +1,11 @@
+package gift.repository;
+
+import gift.entity.ProductEntity;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
+    List<ProductEntity> findByNameContaining(String name);
+}
