+++ src/main/java/gift/util/PasswordUtil.java
@@ -1,0 +1,14 @@
+package gift.util;
+
+import java.util.Base64;
+
+public class PasswordUtil {
+
+    public static String encodePassword(String password) {
+        return Base64.getEncoder().encodeToString(password.getBytes());
+    }
+
+    public static String decodePassword(String encodedPassword) {
+        return new String(Base64.getDecoder().decode(encodedPassword));
+    }
+}
