--- src/main/java/gift/wishlist/domain/WishList.java
@@ -1,106 +1,0 @@
-package gift.wishlist.domain;
-
-public class WishList {
-
-    private Long id;
-    private Long memberId;
-    private Long productId;
-
-    public WishList() {
-
-    }
-
-    public WishList(Long memberId, Long productId) {
-        this.memberId = memberId;
-        this.productId = productId;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getMemberId() {
-        return memberId;
-    }
-
-    public void setMemberId(Long memberId) {
-        this.memberId = memberId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public void setProductId(Long productId) {
-        this.productId = productId;
-    }
-
-    public static class JoinedWishList{
-
-        private Long id;
-        private Long memberId;
-        private Long productId;
-        private String name;
-        private Integer price;
-        private String imageUrl;
-
-        public JoinedWishList() {
-
-        }
-
-        public Long getId() {
-            return id;
-        }
-
-        public void setId(Long id) {
-            this.id = id;
-        }
-
-        public Long getMemberId() {
-            return memberId;
-        }
-
-        public void setMemberId(Long memberId) {
-            this.memberId = memberId;
-        }
-
-        public Long getProductId() {
-            return productId;
-        }
-
-        public void setProductId(Long productId) {
-            this.productId = productId;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public Integer getPrice() {
-            return price;
-        }
-
-        public void setPrice(Integer price) {
-            this.price = price;
-        }
-
-        public String getImageUrl() {
-            return imageUrl;
-        }
-
-        public void setImageUrl(String imageUrl) {
-            this.imageUrl = imageUrl;
-        }
-
-    }
-
-
-}
