--- src/main/java/gift/wishlist/domain/WishListEntity.java
@@ -1,52 +1,0 @@
-package gift.wishlist.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "wish_lists")
-public class WishListEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "member_id", nullable = false)
-    private Long memberId;
-
-    @Column(name = "product_id", nullable = false)
-    private Long productId;
-
-    public WishListEntity() {
-
-    }
-
-    public WishListEntity(Long memberId, Long productId) {
-        this.memberId = memberId;
-        this.productId = productId;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Long getMemberId() {
-        return memberId;
-    }
-
-    public void setMemberId(Long memberId) {
-        this.memberId = memberId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public void setProductId(Long productId) {
-        this.productId = productId;
-    }
-}
