--- src/main/java/gift/wishlist/repository/WishListRepository.java
@@ -1,19 +1,0 @@
-package gift.wishlist.repository;
-
-import gift.wishlist.domain.WishList;
-import gift.wishlist.domain.WishListEntity;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import org.springframework.jdbc.core.BeanPropertyRowMapper;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WishListRepository extends JpaRepository<WishListEntity, Long> {
-
-    List<WishListEntity> findByMemberId(Long memberId);
-
-}
